+++ changelog.md
@@ -38,6 +38,7 @@ Codion Change Log
 - DefaultButtonPanelBuilder.enableTransferFocusOnEnter() bug fixed, was restricted to JButton.
 - ButtonPanelBuilder.fixedButtonSize() and buttonGroup() added.
 - SelectionDialogBuilder.defaultSelection() moved to sub-interfaces.
+- ProgressDialogBuilder.onPublish() added.
 ### is.codion.swing.framework.ui
 - EntityApplicationPanel.Builder.applicationModelFactory() and applicationPanelFactory() renamed applicationModel() and applicationPanel().
 - EntityApplicationPanel.createLogControls() now protected final.